+++ app/client/cypress/locators/SignupPage.json
@@ -2,10 +2,8 @@
   "username": "input[name='email']",
   "password": "input[name='password']",
   "submitBtn": "button[type='submit']",
-  "roleDropdown": ".setup-dropdown:first",
   "proficiencyGroupButton": "[data-testid='t--user-proficiency'] button",
   "useCaseGroupButton": "[data-testid='t--user-use-case'] button",
-  "useCaseDropdown": ".setup-dropdown:last",
   "dropdownOption": ".rc-select-item-option:first",
   "getStartedSubmit": ".t--get-started-button"
 }
