+++ app/client/cypress/locators/welcomePage.json
@@ -4,12 +4,8 @@
   "email": ".t--welcome-form-email input",
   "password": ".t--welcome-form-password input",
   "verifyPassword": ".t--welcome-form-verify-password input",
-  "proficiencyGroupButton": ".radio_group__button_container button",
-  "roleDropdown": ".t--welcome-form-role-dropdown .setup-dropdown",
-  "roleDropdownOption": ".rc-select-item-option",
-  "roleInput": ".t--welcome-form-role-input input",
-  "useCaseDropdown": ".t--welcome-form-role-usecase",
-  "useCaseDropdownOption": ".rc-select-item-option-content",
+  "proficiencyGroupButton": "[data-testid='t--user-proficiency'] button",
+  "useCaseGroupButton": "[data-testid='t--user-use-case'] button",
   "continueButton": ".t--welcome-form-continue-button",
   "submitButton": ".t--welcome-form-submit-button",
   "dataCollection": ".t--welcome-form-datacollection",
