+++ app/client/src/ce/constants/forms.ts
@@ -28,11 +28,8 @@ export const WELCOME_FORM_NAME_FIELD_NAME = "name";
 export const WELCOME_FORM_EMAIL_FIELD_NAME = "email";
 export const WELCOME_FORM_PASSWORD_FIELD_NAME = "password";
 export const WELCOME_FORM_VERIFY_PASSWORD_FIELD_NAME = "verify_password";
-export const WELCOME_FORM_ROLE_FIELD_NAME = "role";
 export const WELCOME_FORM_PROFICIENCY_LEVEL = "proficiency";
-export const WELCOME_FORM_ROLE_NAME_FIELD_NAME = "role_name";
 export const WELCOME_FORM_USECASE_FIELD_NAME = "useCase";
-export const WELCOME_FORM_CUSTOM_USECASE_FIELD_NAME = "custom_useCase";
 
 export const SETTINGS_FORM_NAME = "SettingsForm";
 export const WELCOME_NON_SUPER_FORM_NAME = "WelcomeNonSuperSetupForm";
