+++ app/client/src/pages/setup/SetupForm.tsx
@@ -10,9 +10,7 @@ import {
   WELCOME_FORM_NAME_FIELD_NAME,
   WELCOME_FORM_PASSWORD_FIELD_NAME,
   WELCOME_FORM_VERIFY_PASSWORD_FIELD_NAME,
-  WELCOME_FORM_CUSTOM_USECASE_FIELD_NAME,
-  WELCOME_FORM_ROLE_FIELD_NAME,
-  WELCOME_FORM_ROLE_NAME_FIELD_NAME,
+  WELCOME_FORM_PROFICIENCY_LEVEL,
 } from "@appsmith/constants/forms";
 import type { FormErrors } from "redux-form";
 import { formValueSelector, getFormSyncErrors, reduxForm } from "redux-form";
@@ -22,14 +20,13 @@ import { SUPER_USER_SUBMIT_PATH } from "@appsmith/constants/ApiConstants";
 import { useState } from "react";
 import { isAirgapped } from "@appsmith/utils/airgapHelpers";
 import {
-  WELCOME_FORM_CUSTOM_USE_CASE_ERROR_MESSAGE,
   WELCOME_FORM_USE_CASE_ERROR_MESSAGE,
   WELCOME_FORM_EMAIL_ERROR_MESSAGE,
   createMessage,
   WELCOME_FORM_STRONG_PASSWORD_ERROR_MESSAGE,
   WELCOME_FORM_GENERIC_ERROR_MESSAGE,
   WELCOME_FORM_PASSWORDS_NOT_MATCHING_ERROR_MESSAGE,
-  WELCOME_FORM_ROLE_ERROR_MESSAGE,
+  WELCOME_FORM_PROFICIENCY_ERROR_MESSAGE,
 } from "@appsmith/constants/messages";
 
 const PageWrapper = styled.div`
@@ -59,12 +56,7 @@ export const firstpageValues = [
   "verifyPassword",
 ];
 
-export const secondPageValues = [
-  "role",
-  "useCase",
-  "custom_useCase",
-  "role_name",
-];
+export const secondPageValues = ["proficiency", "useCase"];
 
 const validate = (values: DetailsFormValues) => {
   const errors: DetailsFormValues = {};
@@ -86,22 +78,13 @@ const validate = (values: DetailsFormValues) => {
     );
   }
 
-  if (!values.role) {
-    errors.role_name = createMessage(WELCOME_FORM_ROLE_ERROR_MESSAGE);
-  }
-
-  if (values.role == "other" && !values.role_name) {
-    errors.role_name = createMessage(WELCOME_FORM_ROLE_ERROR_MESSAGE);
+  if (!values.proficiency) {
+    errors.proficiency = createMessage(WELCOME_FORM_PROFICIENCY_ERROR_MESSAGE);
   }
 
   if (!values.useCase) {
     errors.useCase = createMessage(WELCOME_FORM_USE_CASE_ERROR_MESSAGE);
   }
-
-  if (values.useCase === "other" && !values.custom_useCase)
-    errors.custom_useCase = createMessage(
-      WELCOME_FORM_CUSTOM_USE_CASE_ERROR_MESSAGE,
-    );
 
   return errors;
 };
@@ -137,27 +120,20 @@ function SetupForm(props: SetupFormProps) {
       form.appendChild(fullName);
     }
 
-    const roleInput = document.createElement("input");
-    roleInput.type = "text";
-    roleInput.name = "role";
-    roleInput.style.display = "none";
-    if (props.role !== "other") {
-      roleInput.value = props.role as string;
-    } else {
-      roleInput.value = props.role_name as string;
-    }
-    form.appendChild(roleInput);
+    const proficiencyInput = document.createElement("input");
+    proficiencyInput.type = "text";
+    proficiencyInput.name = "proficiency";
+    proficiencyInput.style.display = "none";
+    proficiencyInput.value = props.proficiency as string;
+    form.appendChild(proficiencyInput);
 
     const useCaseInput = document.createElement("input");
     useCaseInput.type = "text";
     useCaseInput.name = "useCase";
     useCaseInput.style.display = "none";
-    if (props.useCase !== "other") {
-      useCaseInput.value = props.useCase as string;
-    } else {
-      useCaseInput.value = props.custom_useCase as string;
-    }
+    useCaseInput.value = props.useCase as string;
     form.appendChild(useCaseInput);
+
     const anonymousDataInput = document.createElement("input");
     anonymousDataInput.type = "checkbox";
     anonymousDataInput.value = isAirgappedFlag ? "false" : "true";
@@ -247,10 +223,8 @@ export default connect((state: AppState) => {
     email: selector(state, WELCOME_FORM_EMAIL_FIELD_NAME),
     password: selector(state, WELCOME_FORM_PASSWORD_FIELD_NAME),
     verify_password: selector(state, WELCOME_FORM_VERIFY_PASSWORD_FIELD_NAME),
-    role: selector(state, WELCOME_FORM_ROLE_FIELD_NAME),
-    role_name: selector(state, WELCOME_FORM_ROLE_NAME_FIELD_NAME),
+    proficiency: selector(state, WELCOME_FORM_PROFICIENCY_LEVEL),
     useCase: selector(state, WELCOME_FORM_USECASE_FIELD_NAME),
-    custom_useCase: selector(state, WELCOME_FORM_CUSTOM_USECASE_FIELD_NAME),
     formSyncErrors: getFormSyncErrors(WELCOME_FORM_NAME)(state),
   };
 }, null)(
