+++ app/client/src/pages/setup/SignupSuccess.tsx
@@ -8,7 +8,7 @@ import { getCurrentUser } from "selectors/usersSelectors";
 import PerformanceTracker, {
   PerformanceTransactionName,
 } from "utils/PerformanceTracker";
-import Landing from "pages/setup/Welcome";
+import UserWelcomeScreen from "pages/setup/UserWelcomeScreen";
 import { Center } from "pages/setup/common";
 import { Spinner } from "design-system";
 import { isValidLicense } from "@appsmith/selectors/tenantSelectors";
@@ -80,7 +80,7 @@ export function SignupSuccess() {
   //TODO(Balaji): Factor in case, where user had closed the tab, while filling the form.And logs back in again.
   if (
     user?.isSuperUser ||
-    (user?.role && user?.useCase) ||
+    ((user?.role || user?.proficiency) && user?.useCase) ||
     shouldEnableFirstTimeUserOnboarding !== "true"
   ) {
     redirectUsingQueryParam();
@@ -91,7 +91,7 @@ export function SignupSuccess() {
       </Center>
     );
   }
-  return <Landing forSuperUser={false} onGetStarted={onGetStarted} />;
+  return <UserWelcomeScreen isSuperUser={false} onGetStarted={onGetStarted} />;
 }
 
 export default requiresAuth(SignupSuccess);
