+++ app/client/src/pages/setup/constants.ts
@@ -3,52 +3,6 @@ export interface OptionType {
   value?: string;
 }
 
-export const roleOptions: OptionType[] = [
-  {
-    label: "Frontend (HTML/JS/React)",
-    value: "frontend engineer",
-  },
-  {
-    label: "Backend (APIs/Databases)",
-    value: "backend engineer",
-  },
-  {
-    label: "Fullstack",
-    value: "fullstack engineer",
-  },
-  {
-    label: "SQL Queries & Basic Coding",
-    value: "business analyst",
-  },
-  {
-    label: "Non Technical",
-    value: "non technical",
-  },
-  {
-    label: "Other",
-    value: "other",
-  },
-];
-
-export const useCaseOptions: OptionType[] = [
-  {
-    label: "Just Exploring",
-    value: "just exploring",
-  },
-  {
-    label: "Personal Project",
-    value: "personal project",
-  },
-  {
-    label: "Work Project",
-    value: "work project",
-  },
-  {
-    label: "Other",
-    value: "other",
-  },
-];
-
 type OptionTypeWithSubtext = OptionType & {
   subtext?: string;
 };
@@ -76,7 +30,7 @@ export const proficiencyOptions: OptionTypeWithSubtext[] = [
   },
 ];
 
-export const useCaseOptionsForNonSuperUser: OptionTypeWithSubtext[] = [
+export const useCaseOptions: OptionTypeWithSubtext[] = [
   {
     label: "Work Project",
     value: "work project",
