+++ app/server/appsmith-server/src/main/java/com/appsmith/server/dtos/UserSignupRequestDTO.java
@@ -19,6 +19,7 @@ public class UserSignupRequestDTO {
 
     private String password;
 
+    @Deprecated
     private String role;
 
     private String proficiency;
